ref: move handleListItemUpdate to controllers.publish.ts
