+++ src/server/components/dashboard/listsItems/item/update/controllers.publish.ts
@@ -1,5 +1,9 @@
 import { Request, Response } from "express";
-import { handleListItemUpdate } from "server/components/dashboard/listsItems/controllers";
+import { EventJsonData, User } from "server/models/types";
+import { logger } from "server/services/logger";
+import { prisma } from "server/models/db/prisma-client";
+import { ListItemEvent } from "@prisma/client";
+import { update } from "server/models/listItem";
 
 export async function listItemUpdateController(req: Request, res: Response): Promise<void> {
   const userId = req.user!.id;
@@ -21,5 +25,49 @@ export async function listItemUpdateController(req: Request, res: Response): Pro
   } catch (error: any) {
     req.flash("errorMsg", `${listItem.jsonData.organisationName} could not be updated. ${error.message}`);
     return res.redirect(listItemUrl);
+  }
+}
+
+export async function handleListItemUpdate(id: number, userId: User["id"]): Promise<void> {
+  logger.info(`${userId} looking for ${id} to update`);
+  const listItem = await prisma.listItem.findUnique({
+    where: { id },
+    include: {
+      history: {
+        orderBy: {
+          time: "desc",
+        },
+      },
+    },
+  });
+
+  if (listItem === null) {
+    logger.error(`${userId} tried to look for ${id}, listItem could not be found`);
+    throw new Error(`Unable to store updates - listItem could not be found`);
+  }
+
+  const editEvent = listItem?.history.find((event) => {
+    // @ts-ignore
+    return event.type === ListItemEvent.EDITED && !!event.jsonData?.updatedJsonData;
+  });
+
+  logger.info(`found edit event ${JSON.stringify(editEvent)}`);
+
+  const auditJsonData: EventJsonData = editEvent?.jsonData as EventJsonData;
+
+  // @ts-ignore
+  if (listItem.jsonData?.updatedJsonData) {
+    await update(id, userId);
+    return;
+  }
+
+  if (auditJsonData?.updatedJsonData) {
+    await update(id, userId);
+    return;
+  }
+
+  if (auditJsonData?.updatedJsonData !== undefined) {
+    // @ts-ignore
+    await update(id, userId, auditJsonData.updatedJsonData);
   }
 }
